requirements.txt
fastapi
uvicorn
python-multipart
replicate
cloudinary
pydantic
shopifyapi
requests
httpx
python-dotenv